requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
celery==5.4.0
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
coverage==7.5.4
cron-descriptor==1.4.3
crontab==1.0.1
Django==5.0.4
django-celery-beat==2.6.0
django-celery-results==2.5.1
django-filter==24.2
django-timezone-field==6.1.0
djangorestframework==3.15.1
drf-yasg==1.21.7
flake8==7.1.0
idna==3.7
inflection==0.5.1
iniconfig==2.0.0
kombu==5.3.7
Markdown==3.6
mccabe==0.7.0
packaging==24.0
pluggy==1.5.0
prompt_toolkit==3.0.47
psycopg2-binary==2.9.9
pycodestyle==2.12.0
pyflakes==3.2.0
pytest==8.2.2
pytest-cov==5.0.0
pytest-django==4.8.0
python-crontab==3.1.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.6
requests==2.32.3
setuptools==70.0.0
six==1.16.0
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
